Grab http JSON data and store this data in PostgreSQL - Scala
These files work together to grab some JSON data from the Federal
Reserve. The data is cleaned and sent to a PostgreSQL db that can be
queried. This is a good example of scraping /acquiring data, parsing
JSON and using JDBC to insert data into Postgres. There is also some
scripting to setup the DB and tables.